mucha OPTIMIZACION nuevo deltaTime (muy extraño) pero ahora funciona a todo tipo de FPS
Se ha conseguido que las animaciones sean estables, es decir misma velocidad de animacion sin importar la velocidad de fps
Se ha realizado bloqueo del personaje si aparece la ventana de ajustes o de salir (no mas bug de dejar una tecla pulsada, hacer clic en salir y ver como el personaje se mueve sin que tu toques nada)

+++ 1EVA/proyecto-eva1/test/PROYECTO_1EVA_RJT/Utilidades/Constantes.cs
@@ -2,7 +2,7 @@
 {
     class Constantes
     {
-        public static int FPS { get; internal set; }
+        public static double FPS { get; internal set; }
         public static bool MUTED = false;
 
         /**
